ml-pharma/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.3.2
azure-cosmos==4.5.1
psutil==5.9.6
prometheus-client==0.19.0
pydantic==2.5.0
python-multipart==0.0.6
plotly==5.17.0
requests==2.31.0
joblib==1.3.2
aiofiles==24.1.0